refactor(profile): move the geo lookups and db-error copy into the module (KAN-415 tail)
The numbered programme steps D1-D9 are all on develop. What is left is the
TAIL: modules that modules.json already declares but whose paths still sit in
src/lib/. `profile` was one of them —

    src/lib/geo/{places-city,places-schools,places-schools-lookup}.ts
                                        -> src/modules/profile/geo/
    src/lib/db-error-copy.ts            -> src/modules/profile/db-error-copy.ts

All four moves are byte-identical renames (places-schools-lookup reads 96% only
because it names its own path in a comment, which this commit rewrites). 12
importers updated across every reference form: the @/ alias, prose in doc
comments, and the derive script.

WHY THIS ONE FIRST

Measured before moving, by running the shipped `is_protected` at both the old
and the new path — the D9 procedure. `geo/` and `db-error-copy.ts` are NOT
founder-owned at either address, so this move carries no KAN-411 ownership
hazard and needs no trailer.

That is not true of the rest of the tail, and the measurement says so plainly:

    src/lib/beta-access/email.ts   FOUNDER-OWNED -> not protected after a move
    src/lib/invite-text.ts         FOUNDER-OWNED -> not protected after a move

Both are NAMED single files in `is_protected`, so unlike D9's directory glob a
move leaves the pattern matching nothing and CTL-035 fails — the hazard is
detectable rather than silent. It still has to be fixed in the same commit as
those moves, and this commit deliberately does not take them on.

ESTATE (KAN-428 Extraction DoD)

  modules.json               the two src/lib entries collapse into the module
                             root; `src/modules/profile/geo/` is NOT listed
                             separately because it is a subdirectory of a path
                             already there, and a redundant entry is one more
                             thing that can fall out of step
  env-access-baseline.json   CTL-037 keys by path, so places-city.ts reads as
                             NEW at the new address and STALE at the old one —
                             two failures from one move, which is the ratchet
                             working. Rekeyed IN PLACE: a first attempt sorted
                             the dict and produced a 62-line diff across 31
                             untouched entries, which is the ensure_ascii
                             collateral-churn mistake in a different costume.
                             The committed diff is one line.
  tests/support/source-paths the `dbErrorCopy` key SURVIVED regeneration
                             without a seed, because the literal was rewritten
                             inside db-error-copy.test.ts as well as in the
                             manifest, so the self-sustaining loop held. Checked
                             empirically rather than assumed — gotcha #31 is the
                             case where it does not.
  kan422-dead-exports.json   left alone. It is in the DoD's ARCHIVE_FILES: a
                             scan taken on a date, where rewriting a path
                             falsifies the record rather than fixing drift.

VERIFICATION

tsc clean; depcruise 0 errors (the 2 pre-existing no-cross-segment-app warnings
are unchanged); no surviving reference to either old path anywhere outside the
archival snapshot; src/lib/geo has 0 tracked files.

One known-noisy local result, stated rather than hidden: two sandbox cases in
tests/scripts/check-env-access.test.js fail while this move is UNCOMMITTED. The
harness clones HEAD for the tree but copies the WORKING-TREE baseline over it,
so it pairs the new baseline against the old paths and the ratchet correctly
reports both halves. Committing makes the pair consistent again. CTL-037 itself
passes in the working tree — 72 reads across 37 files, exact match.

Refs: KAN-415, KAN-428, KAN-411
