+++ tests/support/source-paths.json
@@ -160,7 +160,6 @@
     "oauthConnections": "src/lib/convene/oauth-connections.ts",
     "postEvent": "src/lib/convene/post-event.ts",
     "resolveWidgets": "src/lib/dashboard/resolve-widgets.ts",
-    "dbErrorCopy": "src/lib/db-error-copy.ts",
     "inviteText": "src/lib/invite-text.ts",
     "moderationPolicy": "src/lib/moderation-policy.ts",
     "pipeline": "src/lib/recommender/v2/pipeline.ts",
@@ -184,6 +183,7 @@
     "platformEnv": "src/modules/platform/env.ts",
     "supabaseServer": "src/modules/platform/supabase-server.ts",
     "supabaseService": "src/modules/platform/supabase-service.ts",
+    "dbErrorCopy": "src/modules/profile/db-error-copy.ts",
     "favourites": "src/modules/profile/favourites.ts",
     "profileFields": "src/modules/profile/profile-fields.ts",
     "sectionVisibility": "src/modules/profile/section-visibility.ts",
